python-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.17
chromadb==0.4.22
pandas==2.1.4
openpyxl==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
